backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1

# ─── Google Gemini AI SDK ─────────────────────────────────────────────────────
google-generativeai==0.8.3

# ─── Image Processing ─────────────────────────────────────────────────────────
Pillow==11.0.0

# ─── Data Validation / Serialisation ─────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1

# ─── HTTP / Networking ────────────────────────────────────────────────────────
python-multipart==0.0.18    # Required for FastAPI file uploads
httpx==0.27.2               # Async HTTP client (optional utility)

# ─── Env / Config ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Logging / Utilities ──────────────────────────────────────────────────────
structlog==24.4.0
